separate form from map component
